refactor(apollo): move apollo files
