+++ packages/app/pages/_app.js
@@ -7,7 +7,7 @@ import React from "react";
 import { ApolloProvider } from "react-apollo";
 import ReactPiwik from "react-piwik";
 import "react-tabs/style/react-tabs.css";
-import withApolloClient from "../src/lib/with-apollo-client";
+import { withApolloClient } from "../src/lib/apollo";
 import "../static/css/custom.css";
 import "../static/css/footer.css";
 import "../static/css/hero.css";
