+++ packages/app/src/lib/index.js
@@ -1,0 +1,1 @@
+export * from "./apollo";
